backend/requirements.txt
flask==3.1.2
flask-cors==5.0.0
gunicorn==20.1.0
supabase==2.8.1
openai==1.59.6
anthropic==0.40.0
pydantic==2.9.2
python-dotenv==1.0.1
pillow==11.3.0
pytesseract==0.3.13
pdfminer.six==20231228
pymupdf==1.25.1
numpy==2.1.3
scikit-learn==1.7.1
pandas==2.2.3
apscheduler==3.11.0
psutil==7.0.0
pgvector==0.2.5
psycopg[binary]==3.2.3
httpx==0.27.2
tenacity==9.0.0
pdf2image==1.17.0
google-generativeai==0.8.3
scipy==1.14.1
pytest==8.0.0
